Change translations for date, timestamp and primary_key to numeric
Octo previously treated INTEGER and NUMERIC as equivalent, however
recent changes have changed that equivalency and now the correct
data type needs to be specified.

In VistA dates and timestamps are numerics with a decimal point.
The primary key data type is also converted to a numeric as it is
possible to have decimals in the IEN columns.