[#11] Make the output of _YDBOCTOVISTAM.m idempotently applyable to Octo
Before each CREATE TABLE statement, put a DROP TABLE IF EXISTS statement.

Also, some re-formatting of the comments so that the routine format
complies with the SAC (esp in this case, first line is routine name).

https://www.hardhats.org/tools/SACC_20191101.pdf section 2.2.1.

Also add commit-verify to pipeline, as we never had a pipeline on this
repo before.