[#13] Various TODOs need to be either converted to issues or removed...
Here's the status of various TODOs:
- TODO: change this to use VistA IO utils: I decided that since this
code is only gonna run on GT.M or YottaDB that this is not important. In
terms of VistA SAC correctness, this is not proper, but our clients
don't really care about this in the end. It will take a few hours to
ensure that path parsing is correct, so that's why I am reluctant to
spend that much time for something that is not really needed.
- TODO: Need to figure out how to pull in primary key information and
ensure the columns are in the create table statement: This seems to be a
stale comment, since I far as I can tell there are no missing columns in
the output, nor have any of our clients ever complained of missing
columns.
- TODO: Use a different variable than ERROR here, ERROR does what I
need, but this isn't really an error: Added a new variable ISSUBFILE.
- TODO: Add support for Foreign Keys when implemented in YDBOcto#773.
This is a true TODO, and it has been added.
- TODO: TIMESTAMP and TEXT aren't valid either. From the code, this is
actually taken care of and they are converted to proper types.
- TODO: PRIMARY_KEY and DATE aren't valid either. From the code, this is
actually taken care of and they are converted to proper types.

Misc changes:
- Add a new line between each table in the output so that it's clear to
a human reader where one table ends and another starts.