[#24] Specify AIM Metadata Type 1 for VistA Fileman Tables
AIM supports metadata for two application global types, a default with
metadata depending on whether or not nodes exist at a certain subscript
level, and type=1 nodes for Fileman files where metadata can be required
for certain nodes that don't exist but there are peers at that level.
When Octo is enhanced to use AIM and YDBOcto#791 is implemented, the
VistA DDL generator will need to be enhanced to accommodate it.

See https://gitlab.com/YottaDB/Util/YDBAIM/-/issues/51 for a background.

Due to the update of this DDL creator code, bump version of file up from
1.3 -> 1.4.