[YottaDB/DBMS/YDBOcto#929] Lowercase SQL IDs
As part of YottaDB/DBMS/YDBOcto#929, SQL identifier storage in Octo is
being switched to lowercase to allow compatibility with Postgres clients
(PGAdmin and Ngpsql (used by PowerBI)). More details are in the linked
issue.

This commit changes the references to the object names (used in the
`keys(...)` expression) to be lowercase, as they won't reference an
existing object anymore after YottaDB/DBMS/YDBOcto#929.

Version bumped to v1.9.